+++ src/components/DonutChart/components/Arc/index.ts
@@ -1,0 +1,2 @@
+export {Arc} from './Arc';
+export type {ArcData} from './Arc';