+++ src/components/DonutChart/components/DonutTooltip/index.ts
@@ -1,0 +1,2 @@
+export {DonutTooltip} from './DonutTooltip';
+export type {DonutTooltipProps} from './DonutTooltip';