+++ src/components/DonutChart/components/index.ts
@@ -1,0 +1,3 @@
+export {Arc} from './Arc';
+export type {ArcData} from './Arc';
+export {DonutTooltip} from './DonutTooltip';