+++ src/components/index.ts
@@ -29,3 +29,5 @@ export {VisuallyHiddenRows} from './VisuallyHiddenRows';
 export {LinePreview} from './LinePreview';
 export {Legend} from './Legend';
 export {LinearGradient} from './LinearGradient';
+export {DonutChart} from './DonutChart';
+export type {DonutChartProps} from './DonutChart';